tests/rtas-test.c: fix Apple endian.h include
Recent rtas-test.c changes added a call to 'be32toh', a function
from the header 'endian.h'. This caused QEMU Travus build to break
because in MacOS be32toh is undefined.

This patch makes a conditional code to include the proper header
file if we're compiling in an __APPLE__ env.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>