linux-user/syscall.c: fix trailing whitespaces and style
There are trailing whitespaces in this file that, if removed
by a text editor that refuses to let the traiing whitespace
alone, which happens to be the editor I am using, will cause
checkpatch.pl to warn about styling problems in the resulting
patch. This happens because the trailing whitespace lines
are using a deprecated style.

To keep the intended change I wanted to do (remove unneeded
labels in do_ioctl_blkpg() and do_sendrecvmsg_locked())
trivial, this patch is another trivial change to fix the
trailing whitespaces and the code style errors that
checkpatch.pl warns about. Doing this change beforehand will
keep the next patch focuses just on the label removal changes.

CC: Riku Voipio <riku.voipio@iki.fi>
CC: Laurent Vivier <laurent@vivier.eu>
Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>