ppc/spapr.c: make ibm,max-associativity-domains scale with user input
The ibm,max-associativity-domains is considering that only a single
associativity domain can exist in the same NUMA level. This is true
today because we do not support any type of NUMA distance user
customization, and all nodes are in the same distance to each other.

To enhance NUMA distance support in the pSeries machine we need to
make this limit flexible. This patch rewrites the max-associativity
logic to consider that multiple associativity domains can co-exist
in the same NUMA level.

This is a guest visible change, so let's wrap the existing logic in
a pre_5_2_numa_associativity flag to avoid changing the expected
behavior of older machine types. This same flag will be used in
the following patches for this same purpose.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>