ppc/spapr.c: introduce SpaprMachineClass::numa_assoc_domains
We can't use the input from machine->numa_state->nodes directly
in the pSeries machine because PAPR does not work with raw distance
values, like ACPI SLIT does. We need to determine common
associativity domains, based on similar performance/distance of the
resources, and set these domains in the associativy array that goes
to the FDT of each resource.

To ease the translation between regular ACPI NUMA distance info
to our PAPR dialect, let's create a matrix called numa_assoc_domains
in the SpaprMachineClass. This matrix is initiated with a call to
spapr_init_numa_assoc_domains(), which will read NUMA information
from the MachineClass, apply a heuristic to determine the associativity
domains for each node, then populate numa_assoc_domains accordingly.

Assuming a populated numa_assoc_domains, the ibm,associativity array
of a CPU with vcpu_id index = indexA and node_id = N will be:

ibm,associativity = {0x5,
                     numa_assoc_domains[N][0],
                     numa_assoc_domains[N][1],
                     numa_assoc_domains[N][2],
                     numa_assoc_domains[N][3],
                     indexA}

allowing us to honour user input in the NUMA setup of the pSeries machine.

As with the previous patches, this change does not affect pre 5.2
machine types. To keep the changes under control, we'll plug in
the basic infrastructure first. The actual heuristic to determine
the associativity domains for each NUMA node will come in a follow-up
patch.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>