target/ppc: Add a fake PMC counter that triggers an EBB event
This commit is quite a hack but servers as an initial prototype to deliver
an EBB event in userspace given a PMC (Performance Monitor Counter) counting
an event (for this commit the fake PMC simulates an instruction counter
that is triggered every 4 seconds, so it needs to be yet adjusted to a ns
granularity, which is closer the average latency of a modern CPU), so it
will mimic roughly a PM_RUN_CYC counter.

It's quite hackish because userspace, after setting the EBBHR register
needs to get set too the EBBRR with the magic value '0xbee' in order to
trigger the delivery of an EBB event, in which case the CPU branches to
the address set in EBBHR.

Hence, for the records, the following code triggers the EBB event as
described above:

int main()
{
 asm volatile (

  "li 4, 0xbee;"
  "mtspr 805, 4;"
  "mtspr 804, 4;"
  "b . -12; "

 :
 :
 :
 );
}

It needs to run in a tight loop to maximize the change of getting a 4s
"tick" in userspace, where EBBRR == 0xbee, otherwise, like if in kernel
space, it won't trigger the EBB event, because EBBRR != 0xbee.

I'll refactor this code adding the proper PMC controls, modeling the MMCR0
and the FSCR_EBB registers.

- changed cpu_get_icount to icount_get

 diff --git a/target/ppc/translate.c b/target/ppc/translate.c
 index f322f19ddb..e8ae883d07 100644
 --- a/target/ppc/translate.c
 +++ b/target/ppc/translate.c
 @@ -82,8 +82,8 @@ void mock(void)
 {
   uint64_t insns;

-  insns = cpu_get_icount();
-  printf("cpu_get_icount() = %ld\n", insns);
+  insns = icount_get();
+  printf("icount_get() = %ld\n", insns);
 }

Signed-off-by: Gustavo Romero <gromero@linux.ibm.com>
Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>