spapr_numa: consider user input when defining associativity
This patch puts all the pieces together to finally allow
user input when defining the NUMA topology of the spapr
guest.

spapr_numa_define_associativity_domains() will read the
distance between the nodes from numa_state and determine
the appropriate NUMA level. The domains, which are going
to be processed in ascending order,  are going to be
matched in this specific NUMA level by using the lower
associativity domain present on each. This will create
an heuristic where the associativities of the first nodes
are re-used in new matches, instead of overwriting them with
a new associativity match. This decision is necessary because
neither we, nor the pSeries kernel, supports multiple
associativity domains for each resource, meaning that we have
to decide which associativity takes precedence.

The Linux kernel takes for granted that the associativity
array of node 0 will be filled with zeroes, and the
associativity of all other nodes must take that into account.
This strategy will ensure that node 0 will always have
0 in all its associativity domains.

These decisions have direct impact on how the user will interact
with the NUMA topology, and none of them are perfect. To keep
this commit message no longer than it already is, let's update the
existing documentation in ppc-spapr-numa.rst with more in depth
details and design considerations/drawbacks in the next patch.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>